% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genRaw.R
\name{genRaw}
\alias{genRaw}
\title{Title Generate raw data domains for an oncology trial}
\usage{
genRaw(N)
}
\arguments{
\item{N}{number of patients}
}
\value{
a list of raw data domains.
\describe{
  \item{dm}{demographics domain}
  \item{tu}{tumor assessment domain}
  \item{rs}{tumor response domain}
  \item{ae}{adverse event domain}
  \item{dd}{death detail domain}
  \item{ds}{disposition domain}
  \item{ex}{exposure domain}
}
}
\description{
Title Generate raw data domains for an oncology trial
}
\examples{
# Generate raw oncology domains for 50 patients
domains <- genRaw(N=50)
dm <- domains$dm
tu <- domains$tu
rs <- domains$rs
ae <- domains$ae
dd <- domains$dd
ds <- domains$ds
ex <- domains$ex

}
